Add UseCaseController
